+++ src/components/pages/order/Main/MainRightSide/Admin/AdimPanel/Form.jsx
@@ -2,49 +2,26 @@
 /* eslint-disable react/prop-types */
 import React from "react";
 import styled from "styled-components";
-
-import TextInput from "../../../../../../reusable-ui/TextInput";
-
 import ImagePreview from "./ImagePreview";
-
-import { getInputTextsConfig, getSelectInputConfig } from "./inputConfig";
-import SelectInput from "../../../../../../reusable-ui/SelectInput";
+import { Inputs } from "./Inputs";
 
 const Form = React.forwardRef(
-  ({ product, onSubmit, onChange, onFocus, onBlur, children }, ref) => {
+  ({ product, onSubmit, children, onChange, onFocus, onBlur }, ref) => {
     // state (vide)
 
     // comportements (vide)
-
-    const inputTexts = getInputTextsConfig(product);
-    const inputSelects = getSelectInputConfig(product);
 
     // affichage
     return (
       <FormStyled onSubmit={onSubmit}>
         <ImagePreview imageSource={product.imageSource} title={product.title} />
-        <div className="input-fields">
-          {inputTexts.map((input) => (
-            <TextInput
-              {...input}
-              key={input.id}
-              onChange={onChange}
-              version="minimalist"
-              onFocus={onFocus}
-              onBlur={onBlur}
-              ref={ref && input.name === "title" ? ref : null}
-            />
-          ))}
-          {inputSelects.map((inputSelect) => (
-            <SelectInput
-              {...inputSelect}
-              key={inputSelect.id}
-              onChange={onChange}
-              onFocus={onFocus}
-              onBlur={onBlur}
-            />
-          ))}
-        </div>
+        <Inputs
+          product={product}
+          onChange={onChange}
+          onFocus={onFocus}
+          onBlur={onBlur}
+          ref={ref}
+        />
         <div className="form-footer">{children}</div>
       </FormStyled>
     );
@@ -62,27 +39,6 @@ const FormStyled = styled.form`
   width: 70%;
   grid-column-gap: 20px;
   grid-row-gap: 8px;
-
-  .input-fields {
-    /* background: blue; */
-    grid-area: 1 / 2 / -2 / 3;
-
-    display: grid;
-    grid-template-rows: repeat(3, 1fr);
-    grid-template-columns: repeat(3, 1fr);
-    grid-row-gap: 8px;
-    grid-column-gap: 8px;
-
-    .title {
-      grid-area: 1/1/2/4;
-    }
-    .image-source {
-      grid-area: 2/1/3/4;
-    }
-    .price {
-      grid-area: 3/1/4/2;
-    }
-  }
 
   .form-footer {
     /* background: green; */
